Move ephemeral styles store into separate module
